requirements.txt
pygame-ce>=2.4.0
Pillow>=10.0.0
pytmx
